refactor(streaming): invert drain composition root
